pyproject.toml
[tool.poetry]
name = "parselcli"
version = "1.2.0"
description = "CLI interpreter for xpath/css selectors"
readme = "README.md"
homepage = "https://scrapecrow.com"
repository = "https://github.com/granitosaurus/parsel-cli"
authors = ["granitosaurus <bernardas.alisauskas@pm.me>"]
license = "GPLv3"
packages = [
    { include = "parselcli" },
]
classifiers=[
    'Environment :: Console',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Operating System :: POSIX',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Topic :: Internet',
]


[tool.poetry.dependencies]
python = "^3.7"
parsel = "^1.6.0"
requests-cache = "^0.9.1"
prompt-toolkit = "^3.0.20"
click = "^8.0.1"
Brotli = "^1.0.9"
toml = "^0.10.2"
loguru = "^0.5.3"
rich = "^10.11.0"
beautifulsoup4 = "^4.10.0"
playwright = { version="^1.17.2", optional=true }
pyperclip = "^1.8.2"
nest-asyncio = "^1.5.4"

[tool.poetry.extras]
browser = ["playwright"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
taskipy = "^1.9.0"
black = "^21.9b0"
pylint = "^2.11.1"
flake8 = "^3.9.2"

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38', 'py39']

[tool.poetry.scripts]
parsel = "parselcli.cli:cli"

[tool.taskipy.settings]
use_vars = true

[tool.taskipy.variables]
pkg = "parselcli"

[tool.taskipy.tasks]
test = "pytest tests/"
fmt = "black {pkg}"
check_fmt = "black --check {pkg}"
lint = "pylint {pkg}"
flake = "flake8 {pkg}"
check = "task check_fmt && task flake && task lint"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
